Add new package features

+++ Package.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Linq;
 using System.Text.Json;
+using ComputerUtils.ConsoleUi;
 
 namespace RIFT_Downgrader
 {
@@ -25,11 +26,21 @@ namespace RIFT_Downgrader
         public string actionName { get; set; } = "";
 
         // File copy
-        public string fileInZip { get; set; } = "";
-        public string fileDestination { get; set; } = "";
+        public string fileInZip { get; set; } = null;
+        public string fileDestination { get; set; } = null; // File download, File copy
 
         // File delete
-        public string file { get; set; } = "";
+        public string file { get; set; } = null;
+
+        // File download
+        public string downloadUri { get; set; } = null;
+
+        // Directory copy
+        public string directoryInZip { get; set; } = null;
+        public string directoryDestination { get; set; } = null;
+
+        // Change launch args
+        public string newLaunchArgs { get; set; } = null;
 
         public ActionType GetActionType()
         {
@@ -57,6 +68,18 @@ namespace RIFT_Downgrader
                             case ActionType.FILECOPY:
                                 Logger.Log("Starting FileCopy from " + fileInZip + " to " + dir + fileDestination);
                                 Console.WriteLine("Starting FileCopy from " + fileInZip + " to " + fileDestination);
+                                if (fileInZip == null)
+                                {
+                                    Logger.Log("Missing fileInZip value in package", LoggingType.Warning);
+                                    Console.WriteLine("No file to copy specified");
+                                    return;
+                                }
+                                if (fileDestination == null)
+                                {
+                                    Logger.Log("Missing fileDestination value in package", LoggingType.Warning);
+                                    Console.WriteLine("No fileDestination specified");
+                                    return;
+                                }
                                 try
                                 {
                                     File.WriteAllBytes(dir + fileDestination, package.GetFile(fileInZip));
@@ -66,23 +89,93 @@ namespace RIFT_Downgrader
                                     Console.ForegroundColor = ConsoleColor.Red;
                                     Console.WriteLine("Copying file failed. Does it exist in the package?");
                                     Console.ForegroundColor = ConsoleColor.White;
+                                    return;
                                 }
                                 
+                                break;
+                            case ActionType.DIRECTORYCOPY:
+                                Logger.Log("Starting DirectoryCopy from " + directoryInZip + " to " + dir + directoryDestination);
+                                Console.WriteLine("Starting DirectoryCopy from " + directoryInZip + " to " + directoryDestination);
+                                if (directoryInZip == null)
+                                {
+                                    Logger.Log("Missing directoryInZip value in package", LoggingType.Warning);
+                                    Console.WriteLine("No directory to copy specified");
+                                    return;
+                                }
+                                if (directoryDestination == null)
+                                {
+                                    Logger.Log("Missing directoryDestination value in package", LoggingType.Warning);
+                                    Console.WriteLine("No directoryDestination specified");
+                                    return;
+                                }
+                                try
+                                {
+                                    if (!directoryDestination.EndsWith("\\") && directoryDestination != "") directoryDestination += "\\";
+                                    if (!directoryInZip.EndsWith("/")) directoryInZip += "/";
+                                    Directory.CreateDirectory(dir + directoryDestination);
+
+                                    foreach(string f in package.files)
+                                    {
+                                        if(f.StartsWith(directoryInZip) && !f.EndsWith("/"))
+                                        {
+                                            string dest = dir + directoryDestination + f.Substring(directoryInZip.Length).Replace("/", "\\");
+                                            Logger.Log("Copying " + f + " to " + dest);
+                                            Console.WriteLine("Copying " + f + " to " + dest);
+                                            if (!Directory.Exists(dest.Substring(0, dest.LastIndexOf("\\")))) Directory.CreateDirectory(dest.Substring(0, dest.LastIndexOf("\\")));
+                                            if(package.DoesFileExist(f)) File.WriteAllBytes(dest, package.GetFile(f));
+                                        }
+                                    }
+                                }
+                                catch (Exception e)
+                                {
+                                    Logger.Log("Copying directory failed. Does it exist in the package?\n" + e.ToString(), LoggingType.Warning);
+                                    Console.ForegroundColor = ConsoleColor.Red;
+                                    Console.WriteLine("Copying directory failed. Does it exist in the package?");
+                                    Console.ForegroundColor = ConsoleColor.White;
+                                    return;
+                                }
+
                                 break;
                             case ActionType.FILEDELETE:
                                 Logger.Log("Deleting " + dir + file);
                                 Console.WriteLine("Deleting " + file);
+                                if(file == null)
+                                {
+                                    Logger.Log("Missing file value in package", LoggingType.Warning);
+                                    Console.WriteLine("No file to delete specified");
+                                    return;
+                                }
                                 if(File.Exists(dir + file)) File.Delete(dir + file);
                                 Logger.Log("Deleted");
                                 Console.WriteLine("Deleted");
                                 break;
+                            case ActionType.FILEDOWNLOAD:
+                                Logger.Log("Downloading from " + downloadUri + " to " + dir + fileDestination);
+                                Console.WriteLine("Downloading from " + downloadUri + " to " + fileDestination);
+                                DownloadProgressUI d = new DownloadProgressUI();
+                                d.StartDownload(downloadUri, dir + fileDestination, true, new Dictionary<string, string> { { "User-Agent", "Rift-Downgrader/" + DowngradeManager.updater.version } });
+                                break;
+                            case ActionType.CHANGELAUNCHARGS:
+                                Logger.Log("Changing launch args to " + newLaunchArgs);
+                                Console.WriteLine("Changing launch args to " + newLaunchArgs);
+                                if (newLaunchArgs == null)
+                                {
+                                    Logger.Log("Missing newLaunchArgs", LoggingType.Warning);
+                                    Console.WriteLine("No launch arguments specified");
+                                    return;
+                                }
+                                DowngradeManager.config.apps.FirstOrDefault(x => (x.canonicalName == o.canonicalName || x.id == o.appId) && x.headset == Headset.RIFT).versions.FirstOrDefault(x => o.versions.Contains(x.version) || o.versionCodes.Contains(x.version_code)).extraLaunchArgs += newLaunchArgs + " ";
+                                DowngradeManager.config.Save();
+                                break;
                             default:
-                                Logger.Log("Unknown action type");
+                                Logger.Log("Unknown action type", LoggingType.Warning);
                                 Console.WriteLine("Unknown action type");
                                 return;
                         }
                         Logger.Log("Action completed");
+                        Console.ForegroundColor = ConsoleColor.Green;
                         Console.WriteLine("Action done");
+                        Console.ForegroundColor = ConsoleColor.White;
                         return;
                     }
                 }
@@ -95,7 +188,10 @@ namespace RIFT_Downgrader
     public enum ActionType
     {
         FILECOPY,
-        FILEDELETE
+        FILEDELETE,
+        FILEDOWNLOAD,
+        DIRECTORYCOPY,
+        CHANGELAUNCHARGS
     }
 
     public class PackageOn
@@ -115,7 +211,11 @@ namespace RIFT_Downgrader
         public void Execute()
         {
             Logger.Log("executing " + metadata.packageName + " " + metadata.packageVersion + " by " + metadata.packageAuthor);
-            foreach (PackageAction a in metadata.actions) a.ExecuteAction(this);
+            foreach (PackageAction a in metadata.actions)
+            {
+                a.ExecuteAction(this);
+                Console.WriteLine();
+            }
         }
 
         public static Package LoadPackage(string path)
@@ -137,11 +237,17 @@ namespace RIFT_Downgrader
             return p;
         }
 
+        public bool DoesFileExist(string name)
+        {
+            return ZipFile.OpenRead(path).GetEntry(name) != null;
+        }
+
         public byte[] GetFile(string name)
         {
-            Stream s = ZipFile.OpenRead(path).GetEntry(name).Open();
-            byte[] b = new byte[s.Length];
-            s.Read(b, 0, (int)s.Length);
+            ZipArchiveEntry e = ZipFile.OpenRead(path).GetEntry(name);
+            byte[] b = new byte[e.Length];
+            Stream s = e.Open();
+            s.Read(b, 0, (int)e.Length);
             s.Close();
             return b;
         }
